+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Print a warning when a stage takes over 10 seconds to finish
 * Add `OperationalContext` class for keeping track of complex operations
 * Use new `OperationalContext`-based classes for handling `Datasource` operations, getting rid of callbacks
+* Use `OperationalContext.Schema` for resolving subschemas
 
 ## 1.4.0-alpha.3 (2024-02-25)